Move TUserSumCounterWrapper & TUserSumHistogramWrapper to cpp file
