+++ cloud/storage/core/libs/user_stats/counter/user_counter.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cloud/storage/core/libs/diagnostics/histogram_types.h>
-
 #include <library/cpp/monlib/dynamic_counters/counters.h>
 #include <library/cpp/monlib/metrics/metric_registry.h>
 
@@ -74,57 +72,6 @@ TBuckets GetUsBuckets();
 
 using TBaseDynamicCounters =
     std::pair<NMonitoring::TDynamicCounterPtr, TString>;
-
-class TUserSumCounterWrapper
-    : public IUserCounter
-{
-private:
-    TVector<TIntrusivePtr<NMonitoring::TCounterForPtr>> Counters;
-    NMonitoring::EMetricType Type = NMonitoring::EMetricType::UNKNOWN;
-
-public:
-    explicit TUserSumCounterWrapper(
-        const TVector<TBaseDynamicCounters>& baseCounters);
-
-    NMonitoring::EMetricType GetType() const;
-    void GetType(NMonitoring::IMetricConsumer* consumer) const override;
-
-    void GetValue(
-        TInstant time,
-        NMonitoring::IMetricConsumer* consumer) const override;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
-class TUserSumHistogramWrapper
-    : public IUserCounter
-{
-    using TExplicitHistogramSnapshot = NMonitoring::TExplicitHistogramSnapshot;
-    using EMetricType = NMonitoring::EMetricType;
-
-private:
-    static constexpr size_t IgnoreBucketCount = 10;
-
-    TVector<TIntrusivePtr<NMonitoring::TDynamicCounters>> Counters;
-    TIntrusivePtr<TExplicitHistogramSnapshot> Histogram;
-    const TBuckets Buckets;
-    EMetricType Type = EMetricType::UNKNOWN;
-
-public:
-    explicit TUserSumHistogramWrapper(
-        const TBuckets& buckets,
-        const TVector<TBaseDynamicCounters>& baseCounters);
-
-    void Clear() const;
-
-    void GetType(NMonitoring::IMetricConsumer* consumer) const override;
-
-    void GetValue(
-        TInstant time,
-        NMonitoring::IMetricConsumer* consumer) const override;
-};
-
-////////////////////////////////////////////////////////////////////////////////
 
 void AddUserMetric(
     IUserCounterSupplier& dsc,
